# Apply `disabledModules` recursively