--- a/lib/tests/modules/disable-recursive/bar.nix
+++ b/lib/tests/modules/disable-recursive/bar.nix
@@ -1,0 +1,5 @@
+{
+  imports = [
+    ../declare-enable.nix
+  ];
+}