--- a/lib/tests/modules/disable-recursive/disable-bar.nix
+++ b/lib/tests/modules/disable-recursive/disable-bar.nix
@@ -1,0 +1,7 @@
+{
+
+  disabledModules = [
+    ./bar.nix
+  ];
+
+}