--- a/lib/tests/modules/disable-recursive/disable-foo.nix
+++ b/lib/tests/modules/disable-recursive/disable-foo.nix
@@ -1,0 +1,7 @@
+{
+
+  disabledModules = [
+    ./foo.nix
+  ];
+
+}