--- a/lib/tests/modules/disable-recursive/foo.nix
+++ b/lib/tests/modules/disable-recursive/foo.nix
@@ -1,0 +1,5 @@
+{
+  imports = [
+    ../declare-enable.nix
+  ];
+}