--- a/lib/tests/modules/disable-recursive/main.nix
+++ b/lib/tests/modules/disable-recursive/main.nix
@@ -1,0 +1,8 @@
+{
+  imports = [
+    ./foo.nix
+    ./bar.nix
+  ];
+
+  enable = true;
+}