###### Motivation for this change
If you have a module `foo.nix` as follows:
```nix
{
  imports = [ ./bar.nix ];
}
```

And then set
```nix
{
  disabledModules = [ ./foo.nix ];
}
```

The module `bar.nix` would still be included, even though `foo.nix` which imports it isn't anymore. This is problematic for e.g. `mkRenamedOptionModule`'s, because none of them would be disabled if the module defining them was disabled, meaning all the renamed options would still apply. See https://github.com/NixOS/nixpkgs/pull/61570 for more info on this. This fixes it.

To be able to implement this correctly, I had to completely restructure how the module system collects modules. While a simpler implementation would be possible, it then shows undefined behavior when e.g. two modules `foo.nix` and `bar.nix` both import `qux.nix`, but only one of `foo.nix`/`bar.nix` gets disabled. In the simpler implementation, `qux.nix` would only be imported for one of them, whereas with this more correct implementation it will get imported in both cases.

Ping @roberth @danbst @nbp @rycee @Profpatsch @aanderse 

###### Things done

- [x] Evaluate how much worse performance is and try to improve it
- [x] Write a test for this
- [x] Extended documentation with information on this change