@roberth 2020-01-03
>  it then shows undefined behavior

Nix doesn't have undefined behavior afaik. I think what you mean is that the output depends on details in the input that should not matter much.