@Infinisil 2020-01-04
Evaluating `nix-instantiate nixos/release-combined.nix -A tested` three times I get:

- Before this change: 220.0s (stddev 2.0s)
- After this change: 226.0s (stddev 1.3s)

So this change makes it about 2.7% slower, not too bad. For more stats, see ofborg's result here: https://github.com/NixOS/nixpkgs/pull/76857/checks?check_run_id=373281930