@Infinisil 2020-01-04
I'll try to optimize it a bit further though, I think there's some potential