@Infinisil 2020-01-05
With that change I get:

- Before: 205.2s (stddev 0.7s, 3 samples)
- After: 211.2s (stddev 3.1s, 6 samples)

Not sure why this time the times are a bit faster than last time, the raw values are 210.574, 212.35, 205.356, 210.043, 214.301, 214.593. Not sure what's up with that 205.356 there either. With those means we get a decrease of 2.9%, so it's about the same. But it certainly can't hurt.