@Infinisil 2020-01-07
Moved `loadModule` to the `let in` scope, no need to expose this in `lib`