@Infinisil 2020-01-09
Using the faster with `genericClosure` now