@utdemir 2020-01-10
After this change, I am unable to use store paths in my module imports. For example, something like:

```nix
{
  imports = [ "${sources.home-manager}/nixos" ];
  ...
}
```

now fails to evaluate with:

```
error: the string '/nix/store/an1yy60yq7p6syfxwbwljb8dmxnkk1ls-source/nixos' is not allowed to refer to a store path (such as '/nix/store/an1yy60yq7p6syfxwbwljb8dmxnkk1ls-source')
```

Is this intentional? If so, is there a workaround?