@Infinisil 2020-01-10
Oh thanks for the report. Turns out the changes in https://github.com/NixOS/nixpkgs/pull/76857#discussion_r364627399 are causing this. I'll make a PR to fix it