@Infinisil 2020-01-10
See above PR which fixes this