@Infinisil 2020-01-16
As uncovered by @evanjs on IRC, this introduces a new case of infinite recursion when imports have a cycle in them, e.g. a module `foo.nix` with the contents
```nix
{
  imports = [ ./foo.nix ];
}
```

now throws
```
error: stack overflow (possible infinite recursion)
```

While I do have a fix for this, I don't think this use-case needs to be supported as I don't know of a case where something like this is needed. To fix this, just break the cycle by not importing the same module again.

So you can say that imports now have to be a DAG instead of an arbitrary graph.

Fix is here in case this is needed in the future:

<details>

```diff
diff --git a/lib/modules.nix b/lib/modules.nix
index e2315290ff0..82981dba73b 100644
--- a/lib/modules.nix
+++ b/lib/modules.nix
@@ -139,16 +139,22 @@ rec {
             disabled = concatLists (catAttrs "disabled" modules);
             inherit modules;
           };
-        in parentFile: parentKey: initialModules: args: collectResults (imap1 (n: x:
+        in parentPath: parentFile: parentKey: initialModules: args:
           let
-            module = loadModule args parentFile "${parentKey}:anon-${toString n}" x;
-            collectedImports = collectStructuredModules module._file module.key module.imports args;
+            # All modules this module imports, loaded and without modules that
+            # are anywhere on the parent path to prevent cycles
+            includedModules = filter (module: ! elem module.key parentPath) (imap1 (n:
+              loadModule args parentFile "${parentKey}:anon-${toString n}"
+            ) initialModules);
+          in collectResults (map (module:
+          let
+            collectedImports = collectStructuredModules (parentPath ++ [ module.key ]) module._file module.key module.imports args;
           in {
             key = module.key;
             module = module;
             modules = collectedImports.modules;
             disabled = module.disabledModules ++ collectedImports.disabled;
-          }) initialModules);
+          }) includedModules);
 
       # filterModules :: String -> { disabled, modules } -> [ Module ]
       #
@@ -165,7 +171,7 @@ rec {
         });
 
     in modulesPath: initialModules: args:
-      filterModules modulesPath (collectStructuredModules unknownModule "" initialModules args);
+      filterModules modulesPath (collectStructuredModules [""] unknownModule "" initialModules args);
 
   /* Massage a module into canonical form, that is, a set consisting
      of ‘options’, ‘config’ and ‘imports’ attributes. */
```
</details>