@evanjs 2020-01-16
Just wanted to confirm that a dry build without @Infinisil's patch does reproduce this issue (`error: stack overflow (possible infinite recursion)`), and with the patch, I am able to get a list of packages that would be built!